+++ Proyecto_Farmacia/cDescuento.cpp
@@ -5,7 +5,7 @@
 
 #include "cDescuento.h"
 
-cDescuento::cDescuento(float porcentaje, string obraSocial) {
+cDescuento::cDescuento(double porcentaje, string obraSocial) {
     this->porcentaje = porcentaje;
     this->obraSocial = obraSocial;
 }
@@ -18,7 +18,7 @@ string cDescuento::GET_OBRASOCIAL() {
     return this->obraSocial;
 }
 
-float cDescuento::GET_PORCENTAJE() {
+double cDescuento::GET_PORCENTAJE() {
     return this->porcentaje;
 
 }
