+++ Proyecto_Farmacia/cGlosinas.cpp
@@ -5,7 +5,7 @@
 
 #include "cGolosinas.h"
 
-cGolosinas::cGolosinas(float _precio, string _nombre, int _cantidad, string _marca, eGolosinas _tipoGolosinas) {
+cGolosinas::cGolosinas(double _precio, string _nombre, int _cantidad, string _marca, eGolosinas _tipoGolosinas) {
 
 }
