+++ Proyecto_Farmacia/cOrtopedia.cpp
@@ -5,7 +5,7 @@
 
 #include "cOrtopedia.h"
 
-cOrtopedia::cOrtopedia(float _precio, string _nombre, int _cantidad, string _marca, eOrtopedia _tipoProduct) {
+cOrtopedia::cOrtopedia(double _precio, string _nombre, int _cantidad, string _marca, eOrtopedia _tipoProduct) {
 
 }
